Bullets only work for 5 seconds
I debated this decision a lot to be honest, and I decided it’d be
better if we did it this way. The other way play styles kept changing
and it got difficult to control and didn’t feel like the normal game,
like I wanted it to be. This seemed like the perfect compromise between
removing them entirely and keeping them as they were.